fix(tests): update tests with latest AOS printing formats
