--- aos-process/ant/crypto/digest/init.lua
@@ -1,8 +1,0 @@
-local SHA3 = require(".crypto.digest.sha3")
-
-local digest = {
-	_version = "0.0.1",
-	keccak256 = SHA3.keccak256,
-}
-
-return digest
