--- aos-process/ant/crypto/util/init.lua
@@ -1,8 +1,0 @@
-local Hex = require(".crypto.util.hex")
-
-local util = {
-	_version = "0.0.1",
-	hex = Hex,
-}
-
-return util
