+++ test/utils.mjs
@@ -5,6 +5,7 @@ import {
   AO_LOADER_OPTIONS,
   DEFAULT_HANDLE_OPTIONS,
   ANT_EVAL_OPTIONS,
+  AOS_ANT_WASM,
 } from '../tools/constants.mjs';
 
 /**
@@ -26,5 +27,22 @@ export async function createAosLoader(params) {
 }
 
 export async function createAntAosLoader() {
-  return createAosLoader(ANT_EVAL_OPTIONS);
+  const handle = await AoLoader(AOS_ANT_WASM, AO_LOADER_OPTIONS);
+  // just to get the mem buffer originally
+  const evalRes = await handle(
+    null,
+    {
+      ...DEFAULT_HANDLE_OPTIONS,
+
+      Tags: [{ name: 'Action', value: 'Eval' }],
+      Data: "print('foo')",
+    },
+
+    AO_LOADER_HANDLER_ENV,
+  );
+
+  return {
+    handle,
+    memory: evalRes.Memory,
+  };
 }
